Rename modules and combine `GetData` and `SendData`
