+++ lumisync/devices.py
