+++ tests/random_multicolor_test.py
@@ -1,7 +1,7 @@
 import random
 import time
 
-from lumisync.utils import SendData
+from lumisync import connection, utils
 
 colors = []
 for i in range(5):
@@ -10,13 +10,13 @@ for i in range(5):
     )
 
 print("Attempting to turn on razer mode...")
-SendData.send_razer_on_off(True)
+connection.send_razer_on_off(True)
 print("Attempting to send colors...")
 print(colors)
-SendData.send_razer_data(SendData.convert_colors(colors))
+connection.send_razer_data(utils.convert_colors(colors))
 print("Colors sent!")
 time.sleep(5)
 print(
     "Attempting to turn off razer mode, it might take up to a minute for the lights to refresh..."
 )
-SendData.send_razer_on_off()
+connection.send_razer_on_off()
