+++ tests/transition_test.py
@@ -2,22 +2,17 @@ import time
 
 import colour
 
-from lumisync.utils import SendData
+from lumisync import connection, utils
 
 print("Attempting to turn on razer mode...")
-SendData.send_razer_on_off(True)
+connection.send_razer_on_off(True)
 
-blue = [0, 0, 255]
-green = [0, 255, 0]
-red = [255, 0, 0]
-
-all_colors = [blue, green, red]
-
-colors = [blue]
+colors = [utils.get_color("blue")]
+all_colors = [utils.get_color(name) for name in ["blue", "green", "red"]]
 
 print("Attempting to send colors...")
 print(colors)
-SendData.send_razer_data(SendData.convert_colors(colors))
+connection.send_razer_data(utils.convert_colors(colors))
 print("Colors sent!")
 
 print("Transitioning to next color in all_colors...")
@@ -29,8 +24,8 @@ for color in all_colors:
     color = colour.Color(rgb=(color[0] / 255, color[1] / 255, color[2] / 255))
     range = list(previous.range_to(color, 10))
     for i in range:
-        SendData.send_razer_data(
-            SendData.convert_colors(
+        connection.send_razer_data(
+            utils.convert_colors(
                 [[int(i.red * 255), int(i.green * 255), int(i.blue * 255)]]
             )
         )
@@ -42,4 +37,4 @@ time.sleep(5)
 print(
     "Attempting to turn off razer mode, it might take up to a minute for the lights to refresh..."
 )
-SendData.send_razer_on_off()
+connection.send_razer_on_off()
